ODROID-M1/M1S: core/uclass: fix failure due to wrong device number for SD
Upstream Linux kernel defines the device node for MMC devices as
'dwmmc@fe2b0000' and 'dwmmc@fe2c0000' while the device tree in U-Boot
defines as 'mmc@fe2b0000' and 'mmc@fe2c0000'. This makes to add new
devices and their device number becomes '3' and '4' for the same
hardware.

	=> mmc list
	dwmmc@fe2b0000: 3 (SD)
	dwmmc@fe2c0000: 2
	sdhci@fe310000: 0 (eMMC)Device 'mmc@fe2b0000': seq 1 is in use by 'dwmmc@fe2b0000'

	mmc@fe2b0000: 1

This problem leads to use the device number '3' for SD card that
supposed to be '1', so instead of adding new boot command 'boot_mmc3',
this fix hacks the device node name to be same as the device node name
in U-Boot.

	=> mmc list
	mmc@fe2b0000: 1 (SD)
	mmc@fe2c0000: 2
	sdhci@fe310000: 0 (eMMC)

Signed-off-by: Dongjin Kim <tobetter@gmail.com>
Change-Id: I4b5c1f620ac7a1610773f38595842e6914772013